[NCL-5619] Fix NPE when using versionOverride
